@@ -1,6 +1,6 @@
 {
   "name": "material-ui-bonus",
-  "version": "0.0.14",
+  "version": "0.1.0",
   "description": "A few extra React components built on top of Material-UI",
   "main": "lib/index.js",
   "module": "lib/index.js",
@@ -24,17 +24,18 @@
     "react"
   ],
   "peerDependencies": {
-    "@material-ui/core": "^3.6.1",
-    "@material-ui/icons": "^3.0.1",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3"
+    "@material-ui/core": "^3.9.2",
+    "@material-ui/icons": "^3.0.2",
+    "@material-ui/styles": "^3.0.0-alpha.10",
+    "react": "^16.8.1",
+    "react-dom": "^16.8.1"
   },
   "dependencies": {
     "classnames": "^2.2.6"
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.5",
-    "@babel/core": "^7.1.6",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
     "@babel/preset-react": "^7.0.0"
   }
 }
